@@ -6,7 +6,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "docker:dev": "cross-env NODE_ENV=development COMMAND=\"yarn dev\" docker-compose up --build",
+    "docker:prod": "docker-compose up --build"
   },
   "dependencies": {
     "@fullcalendar/core": "^6.1.15",
@@ -58,14 +60,15 @@
     "next": "14.2.5",
     "next-auth": "^5.0.0-beta.21",
     "prop-types": "^15.8.1",
-    "react": "^18",
+    "react": "^18.2.0",
     "react-big-calendar": "^1.14.1",
     "react-day-picker": "8.10.1",
-    "react-dom": "^18",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.53.0",
     "react-icons": "^5.3.0",
     "react-lottie": "^1.2.4",
+    "sharp": "^0.33.5",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "use-debounce": "^10.0.3",
@@ -77,6 +80,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "cross-env": "^7.0.3",
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
     "postcss": "^8",
